Add vertical and horizontal center parameters
Add horizontal and vertical centers to the function. Horizontal center is useful when writing scripts with output or company branding. Vertical center can be useful when using native powershell progress bars and wanting clean console output without having to calculate the number of lines to add before manually.

Horizontal center will center the output horizontally based on the current size of the powershell window.

Vertical center will start writing the output to the vertical center of the powershell window.